Pushing the "new map" from within my scene. Added lines to the "TiledMap" script to spawn in coins and attach them to the camera (World Controller).
Added lines to the "TiledMap" script to create colliders that extend downwards and prevent the player from falling off the map if they're trying to move into a space with a tile (but they can still fall off the edges/through the gaps).

Added a timer into the camera (World Controller).

Note: for the Keyboard movement, there _is_ a reset button tied to the "R" key, but this merely resets the player's position and not the whole map. Could consider making it reset the scene.